requirements.txt
Flask==2.0.3
Werkzeug==2.2.2
pymongo
python-dotenv
mysql-connector-python
Jinja2
click
itsdangerous
gunicorn
